x/promos/types/message_init_light_client_test.go
package types

import (
	"testing"

	"promos/testutil/sample"

	sdkerrors "github.com/cosmos/cosmos-sdk/types/errors"
	"github.com/stretchr/testify/require"
)

func TestMsgInitLightClient_ValidateBasic(t *testing.T) {
	tests := []struct {
		name string
		msg  MsgInitLightClient
		err  error
	}{
		{
			name: "invalid address",
			msg: MsgInitLightClient{
				Creator: "invalid_address",
			},
			err: sdkerrors.ErrInvalidAddress,
		}, {
			name: "valid address",
			msg: MsgInitLightClient{
				Creator: sample.AccAddress(),
			},
		},
	}
	for _, tt := range tests {
		t.Run(tt.name, func(t *testing.T) {
			err := tt.msg.ValidateBasic()
			if tt.err != nil {
				require.ErrorIs(t, err, tt.err)
				return
			}
			require.NoError(t, err)
		})
	}
}
